CardTranslator has moved package, both normal class and test class, because the dependencies didn't make sense for the previous location. Also added "controller" and "model" packages to tests under "Framework". Just to try and keep it more sorted, now I was doing changes there anyway.
