+++ app/src/test/java/com/example/gruppe9_kabalerobot/Framework/model/FoundationTest.java
@@ -1,4 +1,4 @@
-package com.example.gruppe9_kabalerobot.Framework;
+package com.example.gruppe9_kabalerobot.Framework.model;
 
 import com.example.gruppe9_kabalerobot.Framework.model.Card;
 import com.example.gruppe9_kabalerobot.Framework.model.Foundation;
